requirements.txt
fastapi
uvicorn
python-dotenv
pdfplumber
pytesseract
easyocr
Pillow
PyPDF2
python-multipart
